+++ lib/flutter_tex.dart
@@ -2,5 +2,6 @@ library flutter_tex;
 
 export 'src/models/tex_view_child.dart';
 export 'src/models/tex_view_decoration.dart';
+export 'src/utils/tex_view_rendering_engine.dart';
 export 'src/utils/tex_view_style.dart';
-export 'src/views/flutter_tex_view_controller.dart';
+export 'src/views/platform_interface.dart';