+++ lib/src/utils/tex_view_rendering_engine.dart
@@ -1,0 +1,22 @@
+import 'dart:convert';
+
+class TeXViewRenderingEngine {
+  final String configurations;
+  String _engineName;
+
+  TeXViewRenderingEngine.katex({this.configurations}) {
+    this._engineName = "katex";
+  }
+
+  TeXViewRenderingEngine.mathjax() : configurations = null {
+    this._engineName = "mathjax";
+  }
+
+  String getConfigurations() {
+    return json.encode(configurations ?? "{}");
+  }
+
+  String getEngineName() {
+    return this._engineName;
+  }
+}